requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bcrypt==3.2.2
blinker==1.8.2
boto3==1.34.144
botocore==1.34.144
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==0.8.1
clerk-sdk-python==0.1.0
click==8.1.7
clikit==0.6.2
colorama==0.4.6
crashtest==0.3.1
cryptography==36.0.2
Deprecated==1.2.14
dnspython==2.6.1
dotty-dict==1.3.1
email_validator==2.2.0
exceptionite
Faker==13.16.0
fastapi==0.111.1
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
frozenlist==1.4.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hashids==1.3.1
hfilesize==0.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hupper==1.10.3
hyperframe==6.0.1
idna==3.7
inflection==0.3.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
masonite
masonite-orm==2.22.1
mdurl==0.1.2
mock==5.1.0
modal==0.63.57
multidict==6.0.5
numpy==2.0.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pastel==0.2.1
pendulum==2.1.2
pip-system-certs==4.0
pluggy==1.5.0
protobuf==4.25.3
pycares==4.4.0
pycparser==2.22
pydantic==1.10.17
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.4.0
pylev==1.4.0
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==0.15.0
python-multipart==0.0.9
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
stripe==10.3.0
svix==1.27.0
synchronicity==0.6.7
tenacity==8.5.0
tldextract==2.2.3
toml==0.10.2
typer==0.12.3
types-certifi==2021.10.8.3
types-Deprecated==1.2.9.20240311
types-python-dateutil==2.9.0.20240316
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.22.0
websockets==12.0
whitenoise==5.2.0
wrapt==1.16.0
yarl==1.9.4
